Move BindToObject as a nested class in Binding
No cleanup performed here
